Add SessionCatalogTestBase
