+++ sql/core/src/test/scala/org/apache/spark/sql/connector/DataSourceV2DataFrameSessionCatalogSuite.scala
@@ -19,17 +19,12 @@ package org.apache.spark.sql.connector
 
 import java.util
 
-import org.scalatest.BeforeAndAfter
-
-import org.apache.spark.sql.{DataFrame, QueryTest, SaveMode}
+import org.apache.spark.sql.{DataFrame, SaveMode, SessionCatalogTestBase}
 import org.apache.spark.sql.catalyst.TableIdentifier
 import org.apache.spark.sql.catalyst.analysis.TableAlreadyExistsException
 import org.apache.spark.sql.connector.catalog._
-import org.apache.spark.sql.connector.catalog.CatalogManager.SESSION_CATALOG_NAME
 import org.apache.spark.sql.connector.expressions.Transform
 import org.apache.spark.sql.errors.QueryCompilationErrors
-import org.apache.spark.sql.internal.SQLConf.V2_SESSION_CATALOG_IMPLEMENTATION
-import org.apache.spark.sql.test.SharedSparkSession
 import org.apache.spark.sql.types.StructType
 
 class DataSourceV2DataFrameSessionCatalogSuite
@@ -147,27 +142,7 @@ object InMemoryTableSessionCatalog {
 }
 
 private [connector] trait SessionCatalogTest[T <: Table, Catalog <: TestV2SessionCatalogBase[T]]
-  extends QueryTest
-  with SharedSparkSession
-  with BeforeAndAfter {
-
-  protected def catalog(name: String): CatalogPlugin = {
-    spark.sessionState.catalogManager.catalog(name)
-  }
-
-  protected val v2Format: String = classOf[FakeV2Provider].getName
-
-  protected val catalogClassName: String = classOf[InMemoryTableSessionCatalog].getName
-
-  before {
-    spark.conf.set(V2_SESSION_CATALOG_IMPLEMENTATION.key, catalogClassName)
-  }
-
-  override def afterEach(): Unit = {
-    super.afterEach()
-    catalog(SESSION_CATALOG_NAME).asInstanceOf[Catalog].clearTables()
-    spark.conf.unset(V2_SESSION_CATALOG_IMPLEMENTATION.key)
-  }
+  extends SessionCatalogTestBase[T, Catalog] {
 
   protected def verifyTable(tableName: String, expected: DataFrame): Unit
